Create MainWindow to split up XcodeListView
